Rename CreateRouter by RouterManager
