+++ src/hooks/useRouter.ts
@@ -4,7 +4,7 @@ import {
   IRouterContextStackStates,
   RouterContext,
 } from "../components/Router";
-import { CreateRouter } from "..";
+import { RouterManager } from "..";
 import { Routers } from "../api/routers";
 
 /**
